+++ test-parts/part-c-internal_scripts/internal-main.js
@@ -1,16 +1,12 @@
-const chai = require("chai");
-const expect = require("chai").expect;
-const chaiThings = require('chai-things');
-const sinon = require('sinon');
+const iValidFile = require("./scripts/s-local_valid");
+const iDefineFile = require("./scripts/s-define_api");
+const iRequestFile = require("./scripts/s-request_api");
+const iRequestPathFile = require("./scripts/s-request_api_paths");
+const iDatabaseFile = require("./scripts/s-supported-databases");
+
 
 function coordinateInternal()
 {
-	var iValidFile = require("./scripts/s-local_valid");
-	var iDefineFile = require("./scripts/s-define_api");
-	var iRequestFile = require("./scripts/s-request_api");
-	var iRequestPathFile = require("./scripts/s-request_api_paths");
-	var iDatabaseFile = require("./scripts/s-supported-databases");
-	
 	describe("C - Internal Scripts", function()
 	{
 		iValidFile.callTestLocalValid();
@@ -21,4 +17,7 @@ function coordinateInternal()
 	});
 }
 
-exports.callInternal = coordinateInternal;
+module.exports =
+{
+	callInternal: coordinateInternal
+};